requirements.txt
accelerate
aioboto3
aiobotocore
aiofiles
aiohappyeyeballs
aiohttp
aioitertools
aiosignal
annotated-types
anyio
anytree
argcomplete
asttokens
attrs
autocommand
autograd
backports.tarfile
beartype
beautifulsoup4
boto3
botocore
bottleneck
brotli
certifi
cffi
chardet
charset-normalizer
click
colorama
colorclass
compressed-rtf
contourpy
cryptography
cycler
decorator
distro
docx2txt
easygui
easyprocess
ebcdic
entrypoint2
executing
extract-msg
fastapi
ffmpy
filelock
fonttools
frozenlist
fsspec
gensim
gradio-client
gradio
graspologic-native
graspologic
greenlet
h11
hnswlib
httpcore
httpx
huggingface-hub
hyppo
idna
importlib-metadata
importlib-resources
inflate64
inflect
iniconfig
ipython
jaraco.collections
jaraco.context
jaraco.functools
jaraco.text
jedi
jinja2
jiter
jmespath
joblib
jsonpickle
kiwisolver
lark
lightrag-hku
llvmlite
lxml
markdown-it-py
markupsafe
matplotlib-inline
matplotlib
mdurl
mkl-fft
mkl-random
mkl-service
more-itertools
mpmath
msoffcrypto-tool
multidict
multivolumefile
nano-vectordb
neo4j
nest-asyncio
networkx
numba
numexpr
numpy
olefile
oletools
ollama
openai
oracledb
orjson
packaging
pandas
parso
patsy
pcodedmp
pdfminer.six
pillow
pip
platformdirs
playwright
pluggy
ply
pot
prompt-toolkit
propcache
psutil
pure-eval
pybcj
pycparser
pycryptodomex
pydantic-core
pydantic
pydub
pyee
pygments
pynndescent
pyparsing
pyppmd
pyqt5-sip
pyqt5
pysocks
pytest-base-url
pytest-playwright
pytest
python-dateutil
python-dotenv
python-multipart
python-pptx
python-slugify
pytz
pyvis
pyyaml
pyzstd
red-black-tree-mod
regex
requests
rich
rtfde
ruff
s3transfer
safehttpx
safetensors
scikit-learn
scipy
seaborn
semantic-version
setuptools
shellingham
sip
six
smart-open
sniffio
soupsieve
speechrecognition
stack-data
starlette
statsmodels
sympy
tenacity
text-unidecode
textract-py3
texttable
threadpoolctl
tiktoken
tokenizers
tomli
tomlkit
torch
tornado
tqdm
traitlets
transformers
typeguard
typer
typing-extensions
tzdata
tzlocal
umap-learn
unicodedata2
urllib3
uvicorn
wcwidth
websockets
wheel
win-inet-pton
win-unicode-console
wrapt
xlrd
xlsxwriter
xxhash
yarl
zipp
textract-py3
